reject reconciliation for incompatible cluster type of control plane
